implement BuilderEditor
