+++ src/plugins/builder/BuilderWidget.vue
@@ -1,48 +1,26 @@
 <script lang="ts">
 import { debounce, uid } from 'quasar';
+import { Dialog } from 'quasar';
 import { Component, Watch } from 'vue-property-decorator';
 
 import WidgetBase from '@/components/Widget/WidgetBase';
 
-import BuilderCatalog from './BuilderCatalog.vue';
+import BuilderEditor from './BuilderEditor.vue';
 import { calculateNormalizedFlows } from './calculateFlows';
-import { SQUARE_SIZE, deprecatedTypes } from './getters';
+import { SQUARE_SIZE, defaultLayoutHeight, defaultLayoutWidth, deprecatedTypes } from './getters';
 import specs from './specs';
 import { builderStore } from './store';
-import { BuilderConfig, BuilderLayout, ClickEvent, FlowPart, PartUpdater, PersistentPart, Rect } from './types';
-
-interface DragAction {
-  hide: boolean;
-  part: PersistentPart;
-  x: number;
-  y: number;
-}
-
-interface ToolAction {
-  label: string;
-  value: string;
-  icon: string;
-  cursor?: string;
-  onClick?: (evt: ClickEvent, part: PersistentPart) => void;
-  onPan?: (args: PanArguments, part: PersistentPart) => void;
-}
+import { BuilderConfig, BuilderLayout, FlowPart, PartUpdater, PersistentPart } from './types';
 
 
-@Component({
-  components: {
-    BuilderCatalog,
-  },
-})
+@Component
 export default class BuilderWidget extends WidgetBase {
   $refs!: {
     grid: any;
   }
 
-  formModalOpen: boolean = false;
-  widgetGridRect: Rect | null = null;
-  partState: Record<string, any> = {};
   flowParts: FlowPart[] = [];
-  calculateFlowFunc: Function = () => { };
+  debouncedCalculate: Function = () => { };
 
   get widgetConfig(): BuilderConfig {
     return {
@@ -52,52 +30,41 @@ export default class BuilderWidget extends WidgetBase {
     };
   }
 
-  get layouts(): BuilderLayout[] {
-    return this.widgetConfig.layoutIds
-      .map(builderStore.layoutById);
+  get layout(): BuilderLayout | null {
+    return builderStore.layoutById(this.widgetConfig.currentLayoutId || '');
   }
 
-  get currentLayout(): BuilderLayout | null {
-    return this.layouts
-      .find(s => s && s.id === this.widgetConfig.currentLayoutId) || null;
+  get editorActive(): boolean {
+    return builderStore.editorActive;
   }
 
-  async updateParts(parts: PersistentPart[]) {
+  async saveParts(parts: PersistentPart[]) {
     const asPersistent = (part: PersistentPart | FlowPart) => {
       /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
       const { transitions, flows, ...persistent } = part as FlowPart;
       return persistent;
     };
 
-    if (!this.currentLayout) {
+    if (!this.layout) {
       return;
     }
 
     // first set local value, to avoid jitters caused by the period between action and vueX refresh
-    this.currentLayout.parts = parts.map(asPersistent);
-    await builderStore.saveLayout(this.currentLayout);
-    this.calculateFlowFunc();
+    this.layout.parts = parts.map(asPersistent);
+    await builderStore.saveLayout(this.layout);
+    this.debouncedCalculate();
   }
 
-  async updatePart(part: PersistentPart) {
-    await this.updateParts(this.parts.map(p => (p.id === part.id ? part : p)));
-  }
-
-  async removePart(part: PersistentPart) {
-    await this.updateParts(this.parts.filter(p => p.id !== part.id));
-  }
-
-  gridRect(): Rect {
-    const { x, y, left, right, top, bottom } = this.$refs.grid.getBoundingClientRect();
-    return { x, y, left, right, top, bottom };
+  async savePart(part: PersistentPart) {
+    await this.saveParts(this.parts.map(p => (p.id === part.id ? part : p)));
   }
 
   get parts(): PersistentPart[] {
-    if (!this.currentLayout) {
+    if (!this.layout) {
       return [];
     }
     const sizes: Record<string, number> = {};
-    return this.currentLayout.parts
+    return this.layout.parts
       .map(part => {
         const actual: PersistentPart = {
           id: uid(),
@@ -117,7 +84,7 @@ export default class BuilderWidget extends WidgetBase {
 
   get updater(): PartUpdater {
     return {
-      updatePart: this.updatePart,
+      updatePart: this.savePart,
     };
   }
 
@@ -134,6 +101,20 @@ export default class BuilderWidget extends WidgetBase {
     return SQUARE_SIZE * val;
   }
 
+  startEditor() {
+    Dialog.create({
+      component: BuilderEditor,
+      initialLayout: this.widgetConfig.currentLayoutId,
+    });
+  }
+
+  async calculate() {
+    await this.$nextTick();
+    if (!this.editorActive) {
+      this.flowParts = calculateNormalizedFlows(this.parts);
+    }
+  }
+
   async migrate() {
     const oldParts: PersistentPart[] = (this.widgetConfig as any).parts;
     if (oldParts) {
@@ -141,60 +122,43 @@ export default class BuilderWidget extends WidgetBase {
       await builderStore.createLayout({
         id,
         title: `${this.widget.title} layout`,
+        width: defaultLayoutWidth,
+        height: defaultLayoutHeight,
         parts: oldParts,
       });
       this.widgetConfig.layoutIds.push(id);
       this.widgetConfig.currentLayoutId = id;
       this.$delete(this.widgetConfig, 'parts');
       this.saveConfig(this.widgetConfig);
-      this.calculateFlowFunc();
     }
   }
 
-  mounted() {
-    this.calculateFlowFunc =
-      debounce(
-        () => this.$nextTick(() => this.flowParts = calculateNormalizedFlows(this.parts)),
-        50,
-        false);
-    this.calculateFlowFunc();
+  created() {
     this.migrate();
+    this.debouncedCalculate = debounce(this.calculate, 50, false);
+    this.debouncedCalculate();
   }
 
-  @Watch('widgetConfig', { deep: true })
-  watchLayout(newCfg: BuilderConfig, oldCfg: BuilderConfig) {
-    if (!oldCfg || newCfg.currentLayoutId !== oldCfg.currentLayoutId) {
-      this.calculateFlowFunc();
-    }
+  @Watch('layout')
+  watchLayout() {
+    this.debouncedCalculate();
+  }
+
+  @Watch('editorActive')
+  watchActive() {
+    this.debouncedCalculate();
   }
 }
 </script>
 
 <template>
   <q-card dark class="text-white column">
-    <q-dialog v-model="formModalOpen" no-backdrop-dismiss maximized>
-      <BuilderForm
-        v-if="formModalOpen"
-        :crud="crud"
-        :widget-grid-rect="widgetGridRect"
-        :parts="parts"
-        :flow-parts="flowParts"
-        @parts="updateParts"
-        @part="updatePart"
-        @dirty="calculateFlowFunc"
-        @remove="removePart"
-        @close="formModalOpen = false"
-      />
-    </q-dialog>
-
     <WidgetToolbar :title="widget.title" :subtitle="displayName">
       <q-item-section side>
-        <q-btn-dropdown
-          flat
-          split
-          icon="mdi-pencil"
-          @click="widgetGridRect = gridRect(); formModalOpen = true"
-        >
+        <q-btn unelevated color="primary" label="Editor" @click="startEditor" />
+      </q-item-section>
+      <q-item-section side>
+        <q-btn-dropdown flat split icon="settings" @click="showForm">
           <q-list dark bordered>
             <ExportAction :crud="crud" />
             <WidgetActions :crud="crud" />
@@ -204,7 +168,7 @@ export default class BuilderWidget extends WidgetBase {
     </WidgetToolbar>
 
     <div class="col">
-      <svg v-if="!formModalOpen" ref="grid" class="grid-base">
+      <svg v-if="!editorActive" ref="grid" class="grid-base">
         <g
           v-for="part in flowParts"
           :transform="`translate(${squares(part.x)}, ${squares(part.y)})`"
@@ -212,7 +176,7 @@ export default class BuilderWidget extends WidgetBase {
           :class="{ clickable: isClickable(part), [part.type]: true }"
           @click="interact(part)"
         >
-          <PartWrapper :part="part" @update:part="updatePart" @dirty="calculateFlowFunc" />
+          <PartWrapper :part="part" @update:part="savePart" @dirty="debouncedCalculate" />
         </g>
       </svg>
     </div>
@@ -222,4 +186,3 @@ export default class BuilderWidget extends WidgetBase {
 <style lang="stylus" scoped>
 @import './grid.styl';
 </style>
-
